+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Toolium Changelog
 
 - Refactor to rename assertScreenshot to assert_screenshot and assertFullScreenshot to assert_full_screenshot
 - Add a config property *appium_app_strings* in [Common] section to request app strings before each Appium test
+- Upload the error screenshot to Jira if the test fails
 
 v0.12.1
 -------